+++ wls-gui-wahllokalsystem/src/components/navigation/TheScoresListItems.vue
